Merge branch 'devel' into fix-typos

# Conflicts:
#	DESCRIPTION

+++ tests/testthat/test-cr_disjoin_filter.R
@@ -80,7 +80,7 @@ test_that("Parameter 'found_in_samples' has the correct structure", {
   expect_error(cr_disjoin_filter(
     data = test_data_filtered,
     found_in_samples = NA
-  ),)
+  ), )
   expect_error(cr_disjoin_filter(
     data = test_data_filtered,
     found_in_samples = c(1, 2, 3)
@@ -95,7 +95,7 @@ test_that("Parameter 'found_in_samples' has the correct structure", {
 ### -----------------------------------------------------------------------###
 ##
 test_that("Output data frame is correct", {
-  data <- test_data_disjoin_filter |> 
+  data <- test_data_disjoin_filter |>
     dplyr::mutate(chrom = as.character(chrom))
   ##
   expect_setequal(colnames(data), result_colnames)
@@ -108,7 +108,7 @@ test_that("Output data frame is correct", {
   expect_true(is.numeric(data$end))
   expect_true(is.character(data$sample_name))
   ##
-  expect_identical(nrow(data), as.integer(106))
+  expect_identical(nrow(data), as.integer(113))
   expect_identical(data$start[1], 150)
   ##
   test_counts_left <- test_data_filtered |>
